Few more fixes
